position buffer works, hyhy ;D

+++ src/test/buffers/PositionBufferTest.java
@@ -7,10 +7,10 @@ import static org.testng.Assert.*;
 
 public class PositionBufferTest {
 
-  private static final float[] ZERO_POSITIONS = new float[] {0, 0, 0, 0};
-  private static final float[] FILLED_POSITIONS = new float[] {1, 2, 3, 4};
-  private static final boolean[] ALL_DEALLOCATED = new boolean[] {false, false};
-  private static final boolean[] ALL_ALLOCATED = new boolean[] {true, true};
+  private static final float[] ZERO_POSITIONS = new float[]{0, 0, 0, 0};
+  private static final float[] FILLED_POSITIONS = new float[]{1, 2, 3, 4};
+  private static final boolean[] ALL_DEALLOCATED = new boolean[]{false, false};
+  private static final boolean[] ALL_ALLOCATED = new boolean[]{true, true};
 
   @Test
   public void initialiseEmpty() {
@@ -29,7 +29,7 @@ public class PositionBufferTest {
 
   @Test(expectedExceptions = RuntimeException.class)
   public void initialiseWithInvalidArraySize() {
-    new PositionBuffer(new float[] {1, 2, 3});
+    new PositionBuffer(new float[]{1, 2, 3});
   }
 
   @Test
@@ -37,7 +37,7 @@ public class PositionBufferTest {
     PositionBuffer buffer = new PositionBuffer(2);
     int id = buffer.allocate(1, 2);
     assertEquals(id, 0);
-    assertEquals(buffer.positionsArray(), new float[] {1, 2, 0, 0});
+    assertEquals(buffer.positionsArray(), new float[]{1, 2, 0, 0});
   }
 
   @Test
@@ -45,18 +45,18 @@ public class PositionBufferTest {
     PositionBuffer buffer = new PositionBuffer(2);
     int first = buffer.allocate(1, 2);
     assertEquals(first, 0);
-    assertEquals(buffer.allocationsArray(), new boolean[] {true, false});
+    assertEquals(buffer.allocationsArray(), new boolean[]{true, false});
     int second = buffer.allocate(3, 4);
     assertEquals(second, 1);
-    assertEquals(buffer.allocationsArray(), new boolean[] {true, true});
-    assertEquals(buffer.positionsArray(), new float[] {1, 2, 3, 4});
+    assertEquals(buffer.allocationsArray(), new boolean[]{true, true});
+    assertEquals(buffer.positionsArray(), new float[]{1, 2, 3, 4});
   }
 
   @Test
   public void deallocate() {
     PositionBuffer buffer = new PositionBuffer(FILLED_POSITIONS);
     buffer.deallocate(0);
-    assertEquals(buffer.allocationsArray(), new boolean[] {false, true});
+    assertEquals(buffer.allocationsArray(), new boolean[]{false, true});
     buffer.deallocate(1);
     assertEquals(buffer.allocationsArray(), ALL_DEALLOCATED);
   }
@@ -64,11 +64,10 @@ public class PositionBufferTest {
   @Test
   public void positionGetters() {
     PositionBuffer buffer = new PositionBuffer(FILLED_POSITIONS);
-    assertEquals(buffer.getX(0), 1);
-    assertEquals(buffer.getY(0), 2);
-    assertEquals(buffer.getX(1), 3);
-    assertEquals(buffer.getY(1), 4);
+    assertEquals((int) buffer.getX(0), 1);
+    assertEquals((int) buffer.getY(0), 2);
+    assertEquals((int) buffer.getX(1), 3);
+    assertEquals((int) buffer.getY(1), 4);
   }
-
 
 }
